+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [3.0.0] - 2019-11-29
+### Changed
+- Merged `JpegFormatter`, `GifFormatter` and `PngFormatter` in one `ImageFormatter`.
+
 ### Removed
 - Support for PHP 7.0 and 7.1
 
@@ -95,6 +98,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2016-10-03
 First version
 
+[3.0.0]: https://github.com/middlewares/error-handler/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/middlewares/error-handler/compare/v1.2.0...v2.0.0
 [1.2.0]: https://github.com/middlewares/error-handler/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/middlewares/error-handler/compare/v1.0.0...v1.1.0